Move the section class into it's suitable location.
