+++ osu.Game.Rulesets.Karaoke/Screens/Skin/Config/LyricConfigSection.cs
@@ -1,8 +1,6 @@
 ﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using osu.Game.Rulesets.Karaoke.Edit.Components.Containers;
-
 namespace osu.Game.Rulesets.Karaoke.Screens.Skin.Config;
 
 internal abstract partial class LyricConfigSection : Section
