+++ osu.Game.Rulesets.Karaoke/Screens/Skin/KaraokeSkinEditorScreen.cs
@@ -7,7 +7,6 @@ using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
 using osu.Game.Graphics.Containers;
 using osu.Game.Overlays;
-using osu.Game.Rulesets.Karaoke.Edit.Components.Containers;
 using osu.Game.Rulesets.Karaoke.Screens.Edit;
 using osu.Game.Skinning;
 using osuTK;
