+++ osu.Game.Rulesets.Karaoke/Screens/Skin/Style/StyleScreen.cs
@@ -5,7 +5,6 @@ using System;
 using System.ComponentModel;
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
-using osu.Game.Rulesets.Karaoke.Edit.Components.Containers;
 using osu.Game.Skinning;
 using osuTK;
 using Container = osu.Framework.Graphics.Containers.Container;
