+++ osu.Game.Rulesets.Karaoke/Screens/Skin/Style/StyleSection.cs
@@ -1,8 +1,6 @@
 ﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using osu.Game.Rulesets.Karaoke.Edit.Components.Containers;
-
 namespace osu.Game.Rulesets.Karaoke.Screens.Skin.Style;
 
 internal abstract partial class StyleSection : Section
